requirements.txt
ultralytics>=8.0.0
streamlit>=1.22.0
Pillow>=9.5.0
numpy>=1.24.3
torch>=2.0.0
torchvision>=0.15.0
